Perfected tree placement on rivers, reexported assets for standardization
Perfected tree placement on rivers by reexporting all assets with the -90 degrees on the x axis in blender. Going forward, make sure to use blender as Y up, same as unity. Changed material colors for tiles. Changed the order of angles in the Hex class to reflect the neighbor choosing method.