Regions, counties, hex finding algos, dev panel
Added regions and counties. Still need to implement adjacency. Also need to check if the counties are too big and divide them into more counties accordingly. Created methods to find hexes within a range and spruced up the dev panel to allow for selecting the type of data to display.