Added landmass recognition (contiguity), minor refactors
Data is created to identify what hexes are part of what island. This includes a new Landmass class which is similar to region and county classes. Made a dict for landhexes so I don't have to check sealevel every time; can just pull directly from the landhexes dict. Will need to go through and remove all the unnecessary "is above sea level" checks.