Landmass recognition, failed county balancer, refactoring
Can now recognize different landmasses. Not too many (if any) landmass dependencies at the moment, though. Created a county balancer fucntion, but it sort of decimates the counties and creates noncontiguous counties. The best way to balance the counties is probably to disallow creation of imabalanced counties in the first place. Minor refactoring to replace in-line sea level checking with the prechecked list.